+++ packages/server/src/lib/agentFormationsApply.ts
@@ -2,24 +2,23 @@ import createDebug from 'debug';
 import { db } from 'src/db';
 
 import {
+  applyCreateChange,
+  applyUpdateChange,
+  failFormationOperation,
+} from './agentFormationsApplyHelpers';
+import {
   buildDependencyGraph,
   buildResolvedParamsMap,
   resolveParamExpressions,
   resolveRefs,
   topologicalSort,
 } from './agentFormationsHelpers';
-import {
-  applyCreateResource,
-  applyDeleteResource,
-  applyUpdateResource,
-} from './agentFormationsResourceHandlers';
+import { applyDeleteResource } from './agentFormationsResourceHandlers';
 import type {
   FormationEvent,
   FormationTemplate,
   ResourceDeclaration,
 } from './agentFormationsTypes';
-
-/* eslint-disable max-lines */
 
 const log = createDebug('soat:formations');
 
@@ -80,100 +79,6 @@ export const handleOrphanedDeletes = async (args: {
 };
 
 type ResourceRow = InstanceType<(typeof db)['AgentFormationResource']>;
-
-const applyCreateChange = async (args: {
-  resourceRow: ResourceRow;
-  resourceType: string;
-  resolvedProperties: Record<string, unknown>;
-  projectId: number;
-  logicalId: string;
-  resolvedIds: Map<string, string>;
-  events: FormationEvent[];
-}): Promise<void> => {
-  const {
-    resourceRow,
-    resourceType,
-    resolvedProperties,
-    projectId,
-    logicalId,
-    resolvedIds,
-    events,
-  } = args;
-  const physicalId = await applyCreateResource({
-    resourceType,
-    resolvedProperties,
-    projectId,
-  });
-  resolvedIds.set(logicalId, physicalId);
-  await resourceRow.update({
-    physicalResourceId: physicalId,
-    status: 'created',
-    lastAppliedProperties: resolvedProperties,
-  });
-  events.push({
-    timestamp: new Date().toISOString(),
-    logicalId,
-    resourceType,
-    action: 'create',
-    status: 'succeeded',
-    physicalResourceId: physicalId,
-  });
-};
-
-const applyUpdateChange = async (args: {
-  resourceRow: ResourceRow;
-  existing: ResourceRow & { physicalResourceId: string };
-  resourceType: string;
-  resolvedProperties: Record<string, unknown>;
-  logicalId: string;
-  resolvedIds: Map<string, string>;
-  events: FormationEvent[];
-}): Promise<void> => {
-  const {
-    resourceRow,
-    existing,
-    resourceType,
-    resolvedProperties,
-    logicalId,
-    resolvedIds,
-    events,
-  } = args;
-  const lastProps = (existing.lastAppliedProperties ?? {}) as Record<
-    string,
-    unknown
-  >;
-  const propertiesChanged =
-    JSON.stringify(lastProps) !== JSON.stringify(resolvedProperties);
-  resolvedIds.set(logicalId, existing.physicalResourceId);
-  if (propertiesChanged) {
-    await applyUpdateResource({
-      resourceType,
-      physicalResourceId: existing.physicalResourceId,
-      resolvedProperties,
-    });
-    await resourceRow.update({
-      status: 'updated',
-      lastAppliedProperties: resolvedProperties,
-    });
-    events.push({
-      timestamp: new Date().toISOString(),
-      logicalId,
-      resourceType,
-      action: 'update',
-      status: 'succeeded',
-      physicalResourceId: existing.physicalResourceId,
-    });
-  } else {
-    events.push({
-      timestamp: new Date().toISOString(),
-      logicalId,
-      resourceType,
-      action: 'no-op',
-      status: 'succeeded',
-      physicalResourceId: existing.physicalResourceId,
-    });
-  }
-};
 
 export const processResourceChange = async (args: {
   logicalId: string;
@@ -250,31 +155,6 @@ export const processResourceChange = async (args: {
     });
     throw error;
   }
-};
-
-const failFormationOperation = async (args: {
-  operation: InstanceType<(typeof db)['AgentFormationOperation']>;
-  formation: InstanceType<(typeof db)['AgentFormation']>;
-  events: FormationEvent[];
-  logicalId: string;
-  resourceType: string;
-  action: 'create' | 'update';
-  errorMessage: string;
-}) => {
-  args.events.push({
-    timestamp: new Date().toISOString(),
-    logicalId: args.logicalId,
-    resourceType: args.resourceType,
-    action: args.action,
-    status: 'failed',
-    error: args.errorMessage,
-  });
-  await args.operation.update({
-    status: 'failed',
-    events: args.events,
-    error: { message: args.errorMessage, logicalId: args.logicalId },
-  });
-  await args.formation.update({ status: 'failed' });
 };
 
 /* eslint-disable-next-line max-lines-per-function */
